Compile out legacy code; minor updates

+++ bootloader/dnc-bootloader.h
@@ -123,7 +123,7 @@ checked int read_config_file(const char *filename);
 void udp_open(void);
 void udp_broadcast_state(const void *buf, const size_t len);
 checked int udp_read_state(void *buf, const size_t len, uint32_t *ip);
-void mtrr_range(const uint64_t base, const uint64_t limit, const int type);
+void mtrr_range(const uint64_t base, const uint64_t limit, const uint8_t type);
 void e820_add(const uint64_t base, const uint64_t length, const uint32_t type);
 
 #endif
